Document caching and performance options
Developers familiar with Mason will likely be familiar with these
already; those coming in from the outside may not. They are important to
understand, and frustrating when you don't know what's happening, so a
little documentation here will go a long way.

Plenty of links to Mason docs for people who want to know more.